requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Flask-SQLAlchemy==3.1.1
requests==2.31.0
passlib[argon2]==1.7.4
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg[binary]==3.2.13
pytest==7.4.3

